Fix off-by-one error in miniFAT stream cutoff size
When the stream size is exactly 4096, data should be loaded from FAT
(not from MiniFAT).

https://msdn.microsoft.com/en-us/library/dd941946.aspx

"Mini Stream Cutoff Size (4 bytes): This integer field MUST be set to
0x00001000. This field specifies the maximum size of a user-defined data
stream that is allocated from the mini FAT and mini stream, and that
cutoff is 4,096 bytes. Any user-defined data stream that is greater than
or equal to this cutoff size must be allocated as normal sectors from
the FAT."